concurrent strategies

less useless logs


multiple files "to_change" + "add" reason


less prints


refactoring


no "add" reason -- locate


WIP
