+++ src/at_tools/tools.rs
@@ -40,7 +40,7 @@ pub async fn at_tools_merged_and_filtered(gcx: Arc<ARwLock<GlobalContext>>) -> H
         ("web".to_string(), Arc::new(AMutex::new(Box::new(crate::at_tools::att_web::AttWeb{}) as Box<dyn Tool + Send>))),
         ("cat".to_string(), Arc::new(AMutex::new(Box::new(crate::at_tools::att_cat::AttCat{}) as Box<dyn Tool + Send>))),
         // ("relevant_files".to_string(), Arc::new(AMutex::new(Box::new(crate::at_tools::att_relevant_files::AttRelevantFiles{}) as Box<dyn Tool + Send>))),
-        ("locate".to_string(), Arc::new(AMutex::new(Box::new(crate::at_tools::att_locate::AttLocate{}) as Box<dyn Tool + Send>))),
+        ("locate".to_string(), Arc::new(AMutex::new(Box::new(crate::at_tools::att_locate::att_locate::AttLocate) as Box<dyn Tool + Send>))),
     ]);
 
     let (ast_on, vecdb_on) = {
